change | to , in @PART
changed all occurances of OR operator "|" to the "," according to modulemanager wiki:
"(,<index>)? : Optional index. Again, not available with every option. Not that these indexes are ''with respect to all name matches'' not the total index in the node. I will support negative indexes for running backwards through the list soon. Also * is not yet supported."

Hope thats sufficient